Move PermissionTest to kinto.core.permission.testing
